Feat: Add navigation to lobby screen from player creation, and add challenge system
This commit introduces a navigation from player creation screen once they press the connect button, to the lobby screen, showing available players. And challenge system, allowing players to challenge or receive one, displaying two options, either deny or accept it

refactor: Replace boardState with Game data class, and improve PlayerCreationScreen player handling

The new player creation handling listens to success or failures, also sends back the player Id so we can keep track of the right person throughout the session
Replacing the boardstate in gameViewModel, the Game data class now contains the boardState,
currentPlayerTurn, gameID, gameStatus, player1, and player2. The firestore will be used to display the state of the game in the future together with this change

fix: Resolve player's own name showing on lobby list

This fix corrects the mistake where the lobby list would display the player's own name when it should not, because it does not make sense for them to be able to challenge themselves