Fix: Prevent crash when clicking board squares
The app was crashing with the following error:
java.lang.IllegalArgumentException: Invalid document reference. Document references must have an even number of segments, but games has 1

This occurred when a user clicked on a square on the game board. The cause was that the makeMove function in GameViewModel was
being called with an empty gameId. This led to an attempt to access
Firestore with an invalid document reference.

This commit resolves the issue by:
- Ensuring that the same instance of GameViewModel is used in both LobbyScreen and GameScreen by using by viewModels() in MainActivity. This prevents the creation of multiple instances of GameViewModel with different states.
- Adding a check in makeMove to ensure that the gameId is not empty before attempting to access Firestore. This prevents the crash from occurring even if the gameId is somehow empty.
- Adding logs in initializeGame, listenForChallengeUpdates, acceptChallenge, createGame and gameScreen to help debug the issue.