Fix: Correctly manage player turn, resolve second player's game screen not showing each others moves and prevent excessive recomposition
Refactor: Improve game screen user interface

This commit introduces several fixes to the game screen.

Issues:
- After first player's turn, they could go again and click another square, this results in both of them not being able to make a move on the third turn, because of player1 and player2 being empty strings making mySymbol "O" for the both of them while currentPlayerTurn "X" on the third turn, always making _isMyTurn false.

The fix for this was to initialize the local game data with the one that was created in the firestore, since _game was never initialized with correct values

- Second issue was the second player's device not displaying the first move by the other player

To resolve this issue, was to locally set _game with the data received from the firestore listener whenever the other player made their move

- Third issue was an excessive amount of recompositions

The cause was a loop between the firestore listener that observed the state of a game and the mutableStateOf's. The firestore listening of "listenForChallengeUpdates" noticed many changes to the document resulting in the code inside "listenForChallengeUpdates" being called many times

The fix for this issue was to add a check if the game had already been initialized once for the game screen, and the other was to only check for changes that had just changed/modified right after, not other types according to the challengerId reducing the total snapshots in the for loop, and also reducing initializeGame times called

 The improvement for ui is minor but helps players understand better whose turn it is by instead of displaying X and O, it shows the players names